requirements.txt
asgiref==3.8.1
distro==1.9.0
Django==5.2.3
django-stubs==5.2.1
django-stubs-ext==5.2.1
et_xmlfile==2.0.0
gunicorn==23.0.0
jpype1==1.5.2
numpy==2.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tabula-py==2.10.0
types-PyYAML==6.0.12.20250516
typing_extensions==4.14.0
tzdata==2025.2
whitenoise==6.9.0